+++ include/fluxgl/fluxgl.h
@@ -21,10 +21,13 @@
 #include "ecs/components/audio_listener.h"
 #include "ecs/components/audio_source.h"
 #include "ecs/components/lifetime.h"
+#include "ecs/components/rigidbody.h"
+#include "ecs/components/box_collider.h"
 
 #include "ecs/systems/render_system.h"
 #include "ecs/systems/audio_system.h"
 #include "ecs/systems/lifetime_system.h"
+#include "ecs/systems/physics_system.h"
 
 #include "input/input_manager.h"